+++ source/CleoBase.h
@@ -5,6 +5,7 @@
 #include "CDebug.h"
 #include "CDmaFix.h"
 #include "CGameMenu.h"
+#include "CModLoaderSystem.h"
 #include "CModuleSystem.h"
 #include "CPluginSystem.h"
 #include "CScriptEngine.h"
@@ -35,6 +36,7 @@ namespace CLEO
         CScriptEngine			ScriptEngine;
         CCustomOpcodeSystem		OpcodeSystem;
         CModuleSystem			ModuleSystem;
+        CModLoaderSystem		ModLoader;
         OpcodeInfoDatabase		OpcodeInfoDb;
 
         int saveSlot = -1; // -1 if not loaded from save